Moved CORS fix to __init__.py
