+++ easyrest/__init__.py
@@ -1,4 +1,18 @@
 from pyramid.config import Configurator
+from pyramid.events import NewRequest
+
+
+#CORS fix for react
+def add_cors_headers_response_callback(event):
+    def cors_headers(request, response):
+        response.headers.update({
+        'Access-Control-Allow-Origin': '*',
+        'Access-Control-Allow-Methods': 'POST,GET,DELETE,PUT,OPTIONS',
+        'Access-Control-Allow-Headers': 'Origin, Content-Type, Accept, Authorization',
+        'Access-Control-Allow-Credentials': 'true',
+        'Access-Control-Max-Age': '1728000',
+        })
+    event.request.add_response_callback(cors_headers)
 
 
 def main(global_config, **settings):
@@ -8,5 +22,6 @@ def main(global_config, **settings):
         config.include('.models')
         config.include('pyramid_jinja2')
         config.include('.routes')
+        config.add_subscriber(add_cors_headers_response_callback, NewRequest)
         config.scan()
     return config.make_wsgi_app()
